// idemp-rest-test/testing/src/types/org.bandahealth.idempiere.rest.d.ts
export interface Account extends BaseEntity {
}
// ...
export interface BaseEntity extends BaseMetadata {
    description: string;
    name: string;
    value: string;
}

export interface BaseListResponse<T> {
    results: T[];
    pagingInfo: Paging;
}

export interface BaseMetadata extends Serializable {
    clientId: number;
    orgId: number;
    uuid: string;
    isActive: boolean;
    created: string;
    createdTimestamp: Date;
}
// ...
export interface Charge extends BaseEntity {
    subType: ReferenceList;
    needAdditionalVisitInformation: boolean;
    chargeInformationList: ChargeInformation[];
    account: Account;
    chargeType: ChargeType;
    locked: boolean;
    isLocked: boolean;
}

export interface ChargeInformation extends BaseEntity {
    shouldFillFromPatient: boolean;
    lineNumber: number;
    dataType: ReferenceList;
    values: ChargeInformationValue[];
}

export interface ChargeInformationSuggestion extends BaseEntity {
    shouldFillFromPatient: boolean;
    lineNumber: number;
    dataType: ReferenceList;
    values: ChargeInformationValueSuggestion[];
    subType: ReferenceList;
}

export interface ChargeInformationValue extends BaseEntity {
    lineNumber: number;
}

export interface ChargeInformationValueSuggestion extends BaseEntity {
    lineNumber: number;
}

export interface ChargeType extends BaseEntity {
}
// ...
export interface ReferenceList extends BaseEntity {
}